feat(TokenEnterAmount): Add new Enter Amount component and useEnterAmount hook (#6242)
### Description
1/5 PR for new Enter Amount component. This PR implements a new
component but doesn't use it anywhere. The tests are absent as this
component will be covered with its usages in the follow-up PRs.

This is only done to reduce amount of lines to review per PR as at the
end this whole feature is about 1k new lines and -1k of removed lines.

### Test plan
Omit tests for now but in the follow-up PRs the following test files
will be testing this component:
- `EnterAmount.test.tsx`
- `EarnEnterAmount.test.tsx`
- `SwapScreen.test.tsx`

There will also be some test files that are gonna be fixed alongside.
Please, see the follow-up PRs for more details.

### Related issues

- Relates to RET-1208

### Backwards compatibility
Yes

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)