fix: phone verification ends at CYA screen always (#6266)
### Description

The phone verification flow is hard coded to finish at the onboarding
screens, which is unexpected because it can only be launched from the
settings flow. This was raised at the release testing of the last
release.

This PR makes the final screen in the flow configurable.

### Test plan


https://linear.app/valora/issue/RET-641/[wallet]-user-not-seeing-updated-ota-translations#comment-15a73164

### Related issues

- Fixes RET-1259

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)