fix: jumpstart intro should only be seen once (#6268)
### Description

This PR fixes the jumpstart intro seen screen from being shown more than
once. The jumpstart slice was part of the redux persist blacklist, now
it is removed and the relevant state variables reset on rehydrate.

### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)