fix(analytics): Send redux_store_size event regularly (#6271)
### Description

Currently, we have a mechanism in place to log the user's redux store
size as an analytics event, however due to a bug in the logic, it only
ever gets logged if a user keeps a session open for >24 hours. This PR
updates the logic to ensure that the redux store size will be logged as
soon as it's first available in a new session, and then subsequently
every minute, whenever the full state is serialized.

### Test plan

Unit and manual tested.

### Related issues

- Fixes #[issue number here]

### Backwards compatibility

<!-- Brief explanation of why these changes are/are not backwards
compatible. -->

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [x] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)