fix: incorrect withdraw percentage amount (#6288)
### Description

As the title - this was a copy paste error, restoring it to the
[previous
logic](https://github.com/valora-inc/wallet/blob/06fb27713e6f1a7934ff591c96efc2e9fe684cbc/src/earn/EarnEnterAmount.tsx#L348).

### Test plan

Applying the max amount now matches the max amount displayed in the
withdraw details.


https://github.com/user-attachments/assets/540854a7-9020-4208-930b-e0042c92a025


### Related issues

n/a

### Backwards compatibility

Y

### Network scalability

If a new NetworkId and/or Network are added in the future, the changes
in this PR will:

- [ ] Continue to work without code changes, OR trigger a compilation
error (guaranteeing we find it when a new network is added)